backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.4
    # via openai
aiosignal==1.3.1
    # via aiohttp
amqp==5.1.1
    # via kombu
anyio==3.7.0
    # via httpcore
arabic-reshaper==3.0.0
    # via xhtml2pdf
asgiref==3.7.2
    # via django
asn1crypto==1.5.1
    # via
    #   oscrypto
    #   pyhanko
    #   pyhanko-certvalidator
async-timeout==4.0.2
    # via
    #   aiohttp
    #   redis
attrs==23.1.0
    # via aiohttp
billiard==4.1.0
    # via celery
black==23.3.0
    # via -r requirements.in
celery[redis]==5.3.0
    # via -r requirements.in
certifi==2023.5.7
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pyhanko
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
contourpy==1.1.0
    # via matplotlib
cryptography==41.0.1
    # via
    #   pyhanko
    #   pyhanko-certvalidator
cssselect2==0.7.0
    # via svglib
cycler==0.11.0
    # via matplotlib
django==4.2.1
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.0.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
exceptiongroup==1.1.1
    # via anyio
fonttools==4.40.0
    # via matplotlib
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
html5lib==1.1
    # via xhtml2pdf
httpcore==0.17.2
    # via httpx
httpx==0.24.1
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-resources==5.12.0
    # via matplotlib
kiwisolver==1.4.4
    # via matplotlib
kombu==5.3.0
    # via celery
lxml==4.9.2
    # via svglib
matplotlib==3.7.1
    # via
    #   -r requirements.in
    #   seaborn
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via black
numpy==1.25.0
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   pandas
    #   seaborn
openai==0.27.8
    # via -r requirements.in
oscrypto==1.3.0
    # via pyhanko-certvalidator
packaging==23.1
    # via
    #   black
    #   matplotlib
pandas==2.0.3
    # via
    #   -r requirements.in
    #   seaborn
pathspec==0.11.1
    # via black
pillow==9.5.0
    # via
    #   matplotlib
    #   reportlab
    #   xhtml2pdf
ping3==4.0.4
    # via -r requirements.in
platformdirs==3.5.1
    # via black
prompt-toolkit==3.0.38
    # via click-repl
psutil==5.9.5
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyhanko==0.19.0
    # via xhtml2pdf
pyhanko-certvalidator==0.23.0
    # via
    #   pyhanko
    #   xhtml2pdf
pyparsing==3.1.0
    # via matplotlib
pypdf==3.11.0
    # via xhtml2pdf
pypng==0.20220715.0
    # via qrcode
python-bidi==0.4.2
    # via xhtml2pdf
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   celery
    #   matplotlib
    #   pandas
python-dotenv==1.0.0
    # via
    #   -r cve_prioritizer/cve_prioritizer/requirements.txt
    #   -r requirements.in
python3-nmap==1.6.0
    # via -r requirements.in
pytz==2023.3
    # via
    #   djangorestframework
    #   pandas
pyyaml==6.0
    # via pyhanko
qrcode==7.4.2
    # via pyhanko
redis==4.5.5
    # via celery
reportlab==3.6.13
    # via
    #   svglib
    #   xhtml2pdf
requests==2.31.0
    # via
    #   -r cve_prioritizer/cve_prioritizer/requirements.txt
    #   openai
    #   pyhanko
    #   pyhanko-certvalidator
seaborn==0.12.2
    # via -r requirements.in
simplejson==3.19.1
    # via python3-nmap
six==1.16.0
    # via
    #   click-repl
    #   html5lib
    #   python-bidi
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlparse==0.4.4
    # via django
svglib==1.5.1
    # via xhtml2pdf
termcolor==2.3.0
    # via -r cve_prioritizer/cve_prioritizer/requirements.txt
tinycss2==1.2.1
    # via
    #   cssselect2
    #   svglib
tomli==2.0.1
    # via black
tqdm==4.65.0
    # via openai
typing-extensions==4.6.3
    # via
    #   asgiref
    #   black
    #   kombu
    #   pypdf
    #   qrcode
tzdata==2023.3
    # via
    #   celery
    #   pandas
tzlocal==5.0.1
    # via pyhanko
uritools==4.0.1
    # via pyhanko-certvalidator
urllib3==1.26.6
    # via
    #   -r cve_prioritizer/cve_prioritizer/requirements.txt
    #   requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
xhtml2pdf==0.2.11
    # via -r requirements.in
yarl==1.9.2
    # via aiohttp
zipp==3.15.0
    # via importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
